requirements.txt
pyTelegramBotAPI
python-dotenv
UltimateRobot
Pillow
requests